requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
numpy>=1.24.0
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers>=3.0.1
duckduckgo-search==4.1.1
jinja2>=3.1.2
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn==21.2.0